### 2025-06-09 #8373 @editorialbot
[PRE REVIEW]: SpectralUnmixing: A general Julia package for unmixing spectroscopy data
**Submitting author:** @pgbrodrick (<a href="http://orcid.org/0000-0001-9497-7661">Philip G Brodrick</a>)
**Repository:** https://github.com/emit-sds/SpectralUnmixing.jl
**Branch with paper.md** (empty if default branch): joss
**Version:** v0.2.6
