### 2025-06-12 88368b4
Issue #51 - Clarify doctring of `write_line_results`
