### 2025-06-12 b3c6915
Issue #59 - Install func to deploy driver script
