### 2025-06-12 e9d2c92
Add test for CLI symlink install
