### 2025-06-12 a7e057e
Issue #58 - Add logo
