+++ README.md
@@ -1,5 +1,9 @@
 <h1 align="center">
+<br>
+<a href="https://github.com/emit-sds/SpectralUnmixing.jl"><img src="docs/src/assets/logo.png" alt="SpectralUnmixing.jl" width="200"></a>
+<br>
 SpectralUnmixing.jl
+<br>
 </h1>
 
 
