+++ dev/search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"location":"api/#API-Documentation","page":"API","title":"API Documentation","text":"","category":"section"},{"location":"api/#Endmember-Library-Functions","page":"API","title":"Endmember Library Functions","text":"","category":"section"},{"location":"api/#SpectralUnmixing.SpectralLibrary","page":"API","title":"SpectralUnmixing.SpectralLibrary","text":"mutable struct SpectralLibrary\n\nA structure for managing spectral data from a file, supporting class labels, scaling, and wavelength filtering.\n\n\n\n\n\n","category":"type"},{"location":"api/#SpectralUnmixing.SpectralLibrary-Tuple{String, String}","page":"API","title":"SpectralUnmixing.SpectralLibrary","text":"SpectralLibrary(file_name::String,\n    class_header_name::String,\n    spectral_starting_column::Int64=2,\n    truncate_end_columns::Int64=0,\n    class_valid_keys=nothing,\n    scale_factor=1.0,\n    wavelength_regions_ignore=[0, 440, 1310, 1490, 1770, 2050, 2440, 2880])\n\nConstructor for the SpectralLibrary struct.\n\nArguments\n\nfile_name::String: Endmember Library data file name, CSV-like.\nclass_header_name::String: Column for class labels in data file.\nspectral_starting_column::Int64=2: Column for start of spectral data\ntruncate_end_columns::Int64=0: Number of columns to ignore from end of data.\nclass_valid_keys=nothing: List of valid class labels.\nscale_factor::Float64=1.0: Spectral scaling factor.\nwavelength_regions_ignore=[0,440,1310,1490,1770,2050,2440,2880]: List of wavelength   regions to ignore as beginning/end pairs in list.\n\n\n\n\n\n","category":"method"},{"location":"api/#SpectralUnmixing.load_data!","page":"API","title":"SpectralUnmixing.load_data!","text":"load_data!(library::SpectralLibrary)\n\nLoad and preprocess spectral data associated with the SpectralLibrary object.\n\nLoad spectral data from a CSV-like file library.file_name.\nProcess wavelength ignore regions into paired sets library.wavelength_regions_ignore.\nLoad library.spectra, library.classes, and library.class_valid_keys.\nSort and load library.wavelengths as nanometers and rearrange spectral data accordingly.\nGenerate mask library.good_bands based on library.wavelength_regions_ignore.\n\nReturns\n\nThe updated SpectralLibrary instance with loaded and processed data.\n\n\n\n\n\n","category":"function"},{"location":"api/#SpectralUnmixing.filter_by_class!","page":"API","title":"SpectralUnmixing.filter_by_class!","text":"filter_by_class!(library::SpectralLibrary)\n\nFilter spectra in the SpectralLibrary based on library.class_valid_keys.\n\nUpdate library.spectra and library.classes to include only those matching `library.\n\nclassvalidkeys`.\n\nNotes\n\nThis function modifies the library in place.\nIf library.class_valid_keys is nothing, logs a message indicating that no filtering\n\nwill occur, and exits without making any changes.\n\n\n\n\n\n","category":"function"},{"location":"api/#SpectralUnmixing.read_envi_wavelengths","page":"API","title":"SpectralUnmixing.read_envi_wavelengths","text":"read_envi_wavelengths(filename::String, nm::Bool=true)\n\nRead wavelength data from an ENVI header file with option to convert from microns to nanometers.\n\nIf no wavelengths are found, returns nothing.\n\nNotes\n\nThe function constructs the expected header filename by replacing the extension of the\n\ngiven filename with .hdr.\n\nThis function logs an error if no wavelength data is found in the header file.\n\n\n\n\n\n","category":"function"},{"location":"api/#SpectralUnmixi
[…]
