### 2025-07-10 #8373 @rkurchin
👋 @lamBOOO, @jamesafful and/or @1oly, would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
