### 2025-10-04 #8373 @rkurchin
👋 @dannys4, @sdahdah, and/or @LaurenceDenneulin, would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
