### 2025-10-04 #8373 @dannys4
While I'm a little familiar with the application of uncertainty quantification to remote sensing, unmixing is far from my area of expertise---looking through the example notebook, I'm rather unfamiliar with most of the terms. I can certainly help review Julia and complete the JOSS checklist, but my review would be rather superficial since I'm not part of the relevant community.
