### 2025-10-06 #8373 @LaurenceDenneulin
It is close to my area of expertise so yes, I would gladly review this submission.
