### 2025-10-09 emit-sds/SpectralUnmixing#66 @1oly
Installation guide
Hello,

This is in relation to the review here: https://github.com/openjournals/joss-reviews/issues/9149

I think the instructions given in the README are a little confusing. Just after the `Pkg.add("SpectralUnmixing")` you mention activation and `CLI.install()`, which is not available for users who _added_ the package but for those how clone or `dev`-ed.

For new users I think it would help with a little more guidance on how to use this package and which features are available, e.g., the CLI.

Thanks
