### 2025-11-05 emit-sds/SpectralUnmixing.jl#71 @1oly
CLI calling issues
Hello again,

This is in relation to the review here: https://github.com/openjournals/joss-reviews/issues/9149

Thanks for fixing the install - works great. However, I'm still a little puzzled about how you use the CLI `unmix.jl`. The `CLI.install()` produce a symlink here `~/.julia/bin/unmix.jl`. I also added to PATH, but when I try to use `julia unmix.jl` outside the DEPOT_PATH, I get an error saying there is no file `unmix.jl` in the current path. I suppose it is because Julia looks in the current path for `unmix.jl` instead of the `$PATH`. If the CLI is to be used system-wide, users might need a little more guidance. If you ok with just using `julia unmix.jl` from the SpectralUnmixing repo, then I guess all is good.

I might have misunderstood your description of how to use the CLI command, but would good to get some clarification, either here or in the README, thanks :)
