### 2025-11-05 emit-sds/SpectralUnmixing.jl#71 @vatsal-j
Hi @1oly, yes we're very much okay with users calling `julia unmix.jl` from the SpectralUnmixing root directory, this was always the expectation initially. The CLI.install() procedure is simply a way to put the script in an accessible location if the repo was installed from Julia's package manager, whose directory structure can be a little byzantine. The fact that Julia only looks in the CWD instead of the $PATH was an oversight, we were following an emergent Julia convention seen in other packages, but those use shell wrappers for their CLI scripts. We opted to just update the documentation and instruct users to call `unmix.jl` either from the DEPOT_PATH or the package repo directory depending on the installation method. That update should be pushed soon with an accompanying release, but please proceed by simply calling unmix.jl from it's location, either in the package root directory or the symlink created by `CLI.install()`
