### 2025-11-12 emit-sds/SpectralUnmixing.jl#71 @1oly
Hi @vatsal-j Thanks and sorry for delaying it unnecessary
