### 2025-11-14 e69171b
summary update
