### 2025-11-14 668a7c5
remove paper from base dir
