### 2025-11-14 #9149 @pgbrodrick
Thanks for the feedback @LaurenceDenneulin - I've added a PR [here](https://github.com/emit-sds/SpectralUnmixing.jl/pull/74) to address your request.  Does this satisfy the request?
