### 2025-11-25 #9149 @rkurchin
First, @pgbrodrick please create an archive (e.g. on Zenodo) for the latest version of the software and send the DOI and associated version tag here. The title and author list of the archive should match those for your JOSS paper, and the license listed there should match your software repo.

I'll send some paper copyedits shortly...
