### 2025-11-25 #9149 @rkurchin
Copyediting, etc....
- in summary section, second paragraph: need the signal specifically be a reflectance? I'm not an expert in this field or these methods, but it would seem to me the general principle ought to apply for a variety of spectral signal types, right?
- line 29: remove comma after "flexible"
- line 42: would "under a single API" perhaps be more relevant/descriptive than "into a single codebase" here?
- line 44: "remote-sensing" is operating as a [compound adjective](https://www.grammarly.com/blog/parts-of-speech/compound-adjectives/) here and should hence be hyphenated
- in references, while in general we do want sentence case paper titles, do check for places that capitalization should be {enforced} with curly braces – e.g. in Combe et al., "Mars" should be capitalized
