### 2025-11-30 892a81c
include bib updates
