+++ joss/paper.bib
@@ -1,5 +1,5 @@
 @article{franke2009hierarchical,
-title = {Hierarchical Multiple Endmember Spectral Mixture Analysis (MESMA) of hyperspectral imagery for urban environments},
+title = {Hierarchical Multiple Endmember Spectral Mixture Analysis ({MESMA}) of hyperspectral imagery for urban environments},
 journal = {Remote Sensing of Environment},
 volume = {113},
 number = {8},
@@ -13,7 +13,7 @@ keywords = {Multiple Endmember Spectral Mixture Analysis (MESMA), Hyperspectral 
 }
 
 @article{roberts1998mapping,
-title = {Mapping Chaparral in the Santa Monica Mountains Using Multiple Endmember Spectral Mixture Models},
+title = {Mapping Chaparral in the {Santa Monica} Mountains Using Multiple Endmember Spectral Mixture Models},
 journal = {Remote Sensing of Environment},
 volume = {65},
 number = {3},
@@ -26,7 +26,7 @@ author = {D.A. Roberts and M. Gardner and R. Church and S. Ustin and G. Scheer a
 }
 
 @article{asner2000biogeophysical,
-title = {A Biogeophysical Approach for Automated SWIR Unmixing of Soils and Vegetation},
+title = {A Biogeophysical Approach for Automated {SWIR} Unmixing of Soils and Vegetation},
 journal = {Remote Sensing of Environment},
 volume = {74},
 number = {1},
@@ -70,7 +70,7 @@ author = {Francisco Ochoa and Philip G. Brodrick and Gregory S. Okin and Eyal Be
 }
 
 @article{combe2008analysis,
-  title={Analysis of OMEGA/Mars express data hyperspectral data using a multiple-endmember linear spectral unmixing model (MELSUM): Methodology and first results},
+  title={Analysis of {OMEGA/Mars} express data hyperspectral data using a multiple-endmember linear spectral unmixing model ({MELSUM}): Methodology and first results},
   author={Combe, J-Ph and Le Mou{\'e}lic, St{\'e}phane and Sotin, C and Gendrin, A and Mustard, JF and Le Deit, L and Launeau, P and Bibring, J-P and Gondet, B and Langevin, Y and others},
   journal={Planetary and Space Science},
   volume={56},
