### 2025-12-01 675af8c
stable RNGs implemented; get_sma_permutation added seed
