### 2025-12-01 9c56346
added RNG to project
