+++ Project.toml
@@ -28,26 +28,28 @@ Pluto = "c3e4b0f8-55cb-11ea-2926-15256bba5781"
 Printf = "de0858da-6303-5e67-8744-51eddeeeb8d7"
 PyCall = "438e738f-606a-5dbb-bf0a-cddfbfd45ab0"
 Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
+StableRNGs = "860ef19b-820b-49d6-a774-d7a799459cd3"
 Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 
 [compat]
-julia               = "1.11"
-ArchGDAL            = "0.10.8"
-ArgParse            = "1.2.0"
-CSV                 = "0.10.15"
-Clustering          = "0.15.8"
-Combinatorics       = "1.0.2"
-DataFrames          = "1.7.0"
-Documenter          = "1.9.0"
-EllipsisNotation    = "1.8.0"
-GDAL                = "1.10.0"
-Interpolations      = "0.15.1"
-JuMP                = "1.25.0"
-ModularIndices      = "0.1.0"
-MultivariateStats   = "0.10.3"
-NLopt               = "1.1.3"
-NMF                 = "1.0.3"
-Plots               = "1.40.13"
-Pluto               = "0.20.8"
-PyCall              = "1.96.4"
+ArchGDAL = "0.10.8"
+ArgParse = "1.2.0"
+CSV = "0.10.15"
+Clustering = "0.15.8"
+Combinatorics = "1.0.2"
+DataFrames = "1.7.0"
+Documenter = "1.9.0"
+EllipsisNotation = "1.8.0"
+GDAL = "1.10.0"
+Interpolations = "0.15.1"
+JuMP = "1.25.0"
+ModularIndices = "0.1.0"
+MultivariateStats = "0.10.3"
+NLopt = "1.1.3"
+NMF = "1.0.3"
+Plots = "1.40.13"
+Pluto = "0.20.8"
+PyCall = "1.96.4"
+StableRNGs = "1.0.4"
+julia = "1.11"
