### 2025-12-02 952ec90
added broadband support
