### 2025-12-03 39071d5
added seed variable to get_sma_perm in simulate_pixel
