+++ Project.toml
@@ -33,6 +33,7 @@ Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 
 [compat]
+julia = "1.11"
 ArchGDAL = "0.10.8"
 ArgParse = "1.2.0"
 CSV = "0.10.15"
@@ -52,4 +53,3 @@ Plots = "1.40.13"
 Pluto = "0.20.8"
 PyCall = "1.96.4"
 StableRNGs = "1.0.4"
-julia = "1.11"
