### 2025-12-04 a10cc05
rng tests
