### 2025-12-04 073a10d
Stable RNG enabled
