+++ joss/paper.bib
@@ -48,7 +48,7 @@ author = {Francisco Ochoa and Philip G. Brodrick and Gregory S. Okin and Eyal Be
 }
 
 @inproceedings{emit2020,
-  title={The Earth surface mineral dust source investigation: An Earth science imaging spectroscopy mission},
+  title={The {Earth} surface mineral dust source investigation: An {Earth} science imaging spectroscopy mission},
   author={Green, Robert O and Mahowald, Natalie and Ung, Charlene and Thompson, David R and Bator, Lori and Bennet, Matthew and Bernas, Michael and Blackway, Natalie and Bradley, Christine and Cha, Jeff and others},
   booktitle={2020 IEEE aerospace conference},
   pages={1--15},
@@ -58,7 +58,7 @@ author = {Francisco Ochoa and Philip G. Brodrick and Gregory S. Okin and Eyal Be
 }
 
 @article{yan2010minerals,
-  title={Minerals mapping of the lunar surface with Clementine UVVIS/NIR data based on spectra unmixing method and Hapke model},
+  title={Minerals mapping of the lunar surface with {Clementine} {UVVIS/NIR} data based on spectra unmixing method and {Hapke} model},
   author={Yan, Bokun and Wang, Runsheng and Gan, Fuping and Wang, Zhenchao},
   journal={Icarus},
   volume={208},
