### 2025-12-08 b995c73
added StableRNGs to MESMA calls and SplitLibrary function
