+++ src/EndmemberLibrary.jl
@@ -391,12 +391,13 @@ split_fraction` of the spectra, respectively.
 - The split is random; consecutive calls with the same `library` may yield different
 results.
 """
-function split_library(library::SpectralLibrary, split_fraction::Float64)
+function split_library(library::SpectralLibrary, split_fraction::Float64, seed::Int64)
+    rng = StableRNG(seed)
 
     if !(0 < split_fraction < 1)
         throw(ArgumentError("split_fraction must be between 0 and 1 (exclusive)."))
     end
-    perm = randperm(size(library.spectra)[1])
+    perm = randperm(rng, size(library.spectra)[1])
 
     split_1 = perm[1:Int(round(split_fraction * length(perm)))]
     split_2 = perm[Int(round(split_fraction * length(perm))):end]
