+++ src/SpectralUnmixing.jl
@@ -139,7 +139,7 @@ is a set of indices corresponding to endmembers belonging to one class.
 function get_sma_permutation(class_idx, num_endmembers::Vector{Int64},
     combination_type::String, library_length::Int64, seed::Int64)
 
-    #rng = StableRNG(seed)
+    rng = StableRNG(seed)
 
     if length(num_endmembers) != 1
         throw(ArgumentError("num_endmembers must be a single value"))
@@ -150,8 +150,7 @@ function get_sma_permutation(class_idx, num_endmembers::Vector{Int64},
 
             perm_class_idx = []
             for class_subset in class_idx
-                #push!(perm_class_idx, shuffle(rng, class_subset))
-                push!(perm_class_idx, Random.shuffle(class_subset))
+                push!(perm_class_idx, shuffle(rng, class_subset))
             end
 
             perm = []
@@ -168,7 +167,7 @@ function get_sma_permutation(class_idx, num_endmembers::Vector{Int64},
             end
 
         else
-            perm = randperm(library_length)[1:num_endmembers[1]]
+            perm = randperm(rng, library_length)[1:num_endmembers[1]]
         end
     else
         perm = convert(Vector{Int64}, 1:library_length)
@@ -276,13 +275,11 @@ function unmix_pixel(library::SpectralLibrary, img_dat_input::Array{Float64}, un
     mc_comp_frac = zeros(n_mc, size(library.spectra)[1] + 1)
     scores = zeros(n_mc)
     for mc in 1:n_mc #monte carlo loop
-        #rng = StableRNG(mc)
-        Random.seed!(StableRNG(mc), mc)
-        #Random.seed!(mc)
+        rng = StableRNG(mc)
 
         d = img_dat
         if isnothing(unc_dat) == false
-            d += (rand(size(d)...) .* 2 .- 1) .* unc_dat
+            d += (rand(rng, size(d)...) .* 2 .- 1) .* unc_dat
         end
 
         if occursin("pinv", optimization)
@@ -323,7 +320,7 @@ function unmix_pixel(library::SpectralLibrary, img_dat_input::Array{Float64}, un
             solutions = []
             costs = zeros(size(options)[1]) .+ 1e12
             if max_combinations != -1 && length(options) > max_combinations
-                perm = randperm(length(options))[1:max_combinations]
+                perm = randperm(rng, length(options))[1:max_combinations]
             else
                 perm = convert(Vector{Int64}, 1:length(options))
             end
@@ -486,7 +483,7 @@ function unmix_line(line::Int64, reflectance_file::String, mode::String,
     max_combinations::Int64=-1, optimization="bvls")
 
     #Random.seed!(13)
-    rng = StableRNG(13)
+    rng = StableRNG(13) # what is this used for?
 
     img_dat, unc_dat, good_data = load_line(
         reflectance_file, reflectance_uncertainty_file, line, library.good_bands,
@@ -620,7 +617,7 @@ function simulate_pixel(library::SpectralLibrary, max_components::Int64,
 
     G = library.spectra[perm, :]
 
-    distribution = rand(max_components)
+    distribution = rand(rng, max_components)
     distribution = distribution ./ sum(distribution)
 
     output_distribution = zeros(size(library.spectra)[1])
