### 2025-12-08 e6a3cc4
mc txt file outputs removed
